+++ api/controllers/products.js
@@ -6,6 +6,9 @@ const products = Router();
 products.get("/", get_products);
 products.post("/", get_product_by_id)
 
+products.get("/cart", get_cart_items)
+// products.post("/cart", add_to_cart)
+
 const arr = [
     {
         id: "1",
@@ -27,7 +30,16 @@ const arr = [
     }
 ];
 
+function get_cart_items(req, res) {
+    
+    // const key = req.headers.
+    
+    return
+}
+
 function get_product_by_id(req, res) {
+
+    // req.headers.host  <-- get header host
 
     const { id } = req.body;
