+++ src/css/Cart.css
@@ -1,0 +1,19 @@
+.cart {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.add_count_to_cart * {
+    margin: 10px;
+}
+
+.add_count_to_cart button {
+    border: solid 1px red;
+    background-color: red;
+    color: white;
+    font-size: large;
+    font-weight: bold;
+    border-radius: 15px;
+}
+
