+++ src/pages/Product_Page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const Product_Page = () => {
+  return (
+    <div>Product_Page</div>
+  )
+}
